requirements.txt
lips-benchmark==0.2.7
torch==2.0.1
torch-geometric==2.6.1
tensorflow==2.8.1
